+++ Ghostbait/Ghostbait/Source Files/ObjectManager.cpp
@@ -6,7 +6,7 @@
 ObjectManager::ObjectManager(MemoryManager* _memMan) : memMan(_memMan) {
 	//TODO: Fix this into Initialize
 	poolListCount = 80;
-	poolList = (_Pool_Base*) memMan->RequestMemory((unsigned int)poolListCount, sizeof(Pool<size_t>));
+	poolList = (_Pool_Base*) memMan->RequestMemory(poolListCount, sizeof(Pool<size_t>));
 	//objectpool.reserve(prefabCount);
 
 	EngineStructure::Update += [=]() {this->Update(); };
